Move targets types to separate "Install" folder
Fixes #4331
