+++ src/ConnectedMode/Install/SonarLintTargets.xml
